Move examples into separate package
This allows running unit tests in root package without executing examples code (which are actually integration tests)
