+++ examples/example_client_test.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2020 ArangoDB GmbH, Cologne, Germany
+// Copyright 2017-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,61 +17,35 @@
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
-// Author Adam Janikowski
-//
 
-package driver_test
+//go:build !auth
+
+package examples
 
 import (
-	"context"
+	"fmt"
 	"log"
-	"os"
-	"testing"
-	"time"
 
 	driver "github.com/arangodb/go-driver"
 	"github.com/arangodb/go-driver/http"
 )
 
-// TestMain creates a simple connection and waits for the server to be ready.
-// This avoid a lot of clutter code in the examples.
-func TestMain(m *testing.M) {
-	// Wait for database connection to be ready.
+func ExampleNewClient() {
+	// Create an HTTP connection to the database
 	conn, err := http.NewConnection(http.ConnectionConfig{
 		Endpoints: []string{"http://localhost:8529"},
 	})
 	if err != nil {
 		log.Fatalf("Failed to create HTTP connection: %v", err)
 	}
+	// Create a client
 	c, err := driver.NewClient(driver.ClientConfig{
 		Connection: conn,
 	})
-
-	waitUntilServerAvailable(context.Background(), c)
-
-	os.Exit(m.Run())
-}
-
-// waitUntilServerAvailable keeps waiting until the server/cluster that the client is addressing is available.
-func waitUntilServerAvailable(ctx context.Context, c driver.Client) bool {
-	instanceUp := make(chan bool)
-	go func() {
-		for {
-			verCtx, cancel := context.WithTimeout(ctx, time.Second*5)
-			if _, err := c.Version(verCtx); err == nil {
-				cancel()
-				instanceUp <- true
-				return
-			} else {
-				cancel()
-				time.Sleep(time.Second)
-			}
-		}
-	}()
-	select {
-	case up := <-instanceUp:
-		return up
-	case <-ctx.Done():
-		return false
+	// Ask the version of the server
+	versionInfo, err := c.Version(nil)
+	if err != nil {
+		log.Fatalf("Failed to get version info: %v", err)
 	}
+	fmt.Printf("Database has version '%s' and license '%s'\n", versionInfo.Version, versionInfo.License)
 }
