+++ examples/main_test.go
@@ -1,0 +1,75 @@
+//
+// DISCLAIMER
+//
+// Copyright 2020-2023 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+
+package examples
+
+import (
+	"context"
+	"log"
+	"os"
+	"testing"
+	"time"
+
+	"github.com/arangodb/go-driver"
+	"github.com/arangodb/go-driver/http"
+)
+
+// TestMain creates a simple connection and waits for the server to be ready.
+// This avoid a lot of clutter code in the examples.
+func TestMain(m *testing.M) {
+	// Wait for database connection to be ready.
+	conn, err := http.NewConnection(http.ConnectionConfig{
+		Endpoints: []string{"http://localhost:8529"},
+	})
+	if err != nil {
+		log.Fatalf("Failed to create HTTP connection: %v", err)
+	}
+	c, err := driver.NewClient(driver.ClientConfig{
+		Connection: conn,
+	})
+
+	waitUntilServerAvailable(context.Background(), c)
+
+	os.Exit(m.Run())
+}
+
+// waitUntilServerAvailable keeps waiting until the server/cluster that the client is addressing is available.
+func waitUntilServerAvailable(ctx context.Context, c driver.Client) bool {
+	instanceUp := make(chan bool)
+	go func() {
+		for {
+			verCtx, cancel := context.WithTimeout(ctx, time.Second*5)
+			if _, err := c.Version(verCtx); err == nil {
+				cancel()
+				instanceUp <- true
+				return
+			} else {
+				cancel()
+				time.Sleep(time.Second)
+			}
+		}
+	}()
+	select {
+	case up := <-instanceUp:
+		return up
+	case <-ctx.Done():
+		return false
+	}
+}
